station travel logic and eus distance favorite , station item

+++ app/src/main/java/com/example/a25c992e3a3b6ce9eeb00901b1988403d/model/SpaceStationItem.kt
@@ -2,10 +2,10 @@ package com.example.a25c992e3a3b6ce9eeb00901b1988403d.model
 
 
 data class SpaceStationItem(
-    val capacity: Long = 0,
+    val capacity: Int = 0,
     val coordinateX: Double = 0.0,
     val coordinateY: Double = 0.0,
     val name: String = "",
-    val need: Long = 0,
-    val stock: Long = 0
+    val need: Int = 0,
+    val stock: Int = 0
 )
